UI Improvements
Fixed alignment of the main menu welcome text by making the max groups picker narrower (it was wider than it needed to be).
The game timer bar is now green while loading to match the fact that it's green when the time remaining is more than 10 seconds.
The game timer bar no longer has a visible track. Only the progress part is colored now.
On the "time's up" screen, "Reset Score" is now "Main Menu".